Reloop/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
scikit-learn
websockets
python-dotenv
python-multipart
bcrypt
thefuzz
python-Levenshtein
